# Improve TPC-DS scan performance