--- a/extensions/spark/kyuubi-spark-connector-tpcds/benchmarks/TPCDSTableGenerateBenchmark-results.txt
+++ b/extensions/spark/kyuubi-spark-connector-tpcds/benchmarks/TPCDSTableGenerateBenchmark-results.txt
@@ -1,15 +1,15 @@
-Java HotSpot(TM) 64-Bit Server VM 1.8.0_311-b11 on Windows 10 10.0
-AMD64 Family 23 Model 96 Stepping 1, AuthenticAMD
+OpenJDK 64-Bit Server VM 1.8.0_332-b09 on Mac OS X 12.4
+Apple M1 Pro
 TPCDS table generates 1000000 rows benchmark:  Best Time(ms)   Avg Time(ms)   Stdev(ms)    Rate(M/s)   Per Row(ns)   Relative
 ----------------------------------------------------------------------------------------------------------------------------
-catalog_returns benchmark                             13120          13143          33          0.1       13120.1       1.0X
-catalog_sales benchmark                               12008          12054          44          0.1       12008.4       1.1X
-customer benchmark                                     7499           7529          28          0.1        7499.2       1.7X
-customer_address benchmark                             4939           4958          16          0.2        4939.1       2.7X
-customer_demographics benchmark                         857            860           5          1.2         856.7      15.3X
-inventory benchmark                                     527            529           2          1.9         527.0      24.9X
-store_returns benchmark                               11643          11703          53          0.1       11643.0       1.1X
-store_sales benchmark                                  9418           9517          91          0.1        9418.2       1.4X
-web_returns benchmark                                 13560          13580          23          0.1       13559.7       1.0X
-web_sales benchmark                                   12635          12748         138          0.1       12634.9       1.0X
+catalog_returns benchmark                             13496          13836         398          0.1       13496.3       1.0X
+catalog_sales benchmark                               10084          10119          56          0.1       10084.0       1.3X
+customer benchmark                                     8900           8917          18          0.1        8900.0       1.5X
+customer_address benchmark                             5287           5295          10          0.2        5287.2       2.6X
+customer_demographics benchmark                         880            888           6          1.1         880.3      15.3X
+inventory benchmark                                     523            550          34          1.9         522.6      25.8X
+store_returns benchmark                               10853          10876          40          0.1       10852.9       1.2X
+store_sales benchmark                                  7764           7818          50          0.1        7764.0       1.7X
+web_returns benchmark                                 12840          12863          23          0.1       12839.7       1.1X
+web_sales benchmark                                   10810          10955         193          0.1       10809.6       1.2X