--- a/extensions/spark/kyuubi-spark-connector-tpcds/src/main/scala/org/apache/kyuubi/spark/connector/tpcds/TPCDSBatchScan.scala
+++ b/extensions/spark/kyuubi-spark-connector-tpcds/src/main/scala/org/apache/kyuubi/spark/connector/tpcds/TPCDSBatchScan.scala
@@ -114,7 +114,7 @@ class TPCDSPartitionReader(
         }
         i += 1
       }
-      InternalRow(reusedRow)
+      InternalRow(reusedRow: _*)
     }
 
   private var currentRow: InternalRow = _