@pan3793 2022-07-01
And I'd like to say, AMD YES 💯